back-end/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
reportlab==4.0.7
numpy>=2.0.0
matplotlib>=3.9.0
Pillow==10.4.0
drf-nested-routers==0.93.4
shcheck==1.6.7
trufflehog3==3.0.10